Move include/tkey/assert.h to include/assert.h and fix references
