+++ libcommon/udiv.c
@@ -1,11 +1,7 @@
 // SPDX-FileCopyrightText: 2025 Glasklar Teknik AB <glasklarteknik.se>
 // SPDX-License-Identifier: BSD-2-Clause
 #include <stdint.h>
-#ifdef SELF_TEST
 #include <assert.h>
-#else
-#include <tkey/assert.h>
-#endif
 
 _Static_assert(sizeof(unsigned int) == 4, "unsupported int size");
